Add s390x arch support (#36034)
* Add s390x drone pipelines

* Increase kustomize version to v4.4.1

* Install golang 1.16 using zypper

* Increase loglevel version to v0.1.4

* Add rancher system agent for s390x

* Add unsupported arch env for s390x

* Increase telemetry version

* Add tini for s390x

* Skip integration tests for s390x

* Skip helm v2 instalation since s390x is not supported

* Skip k3s installation for s390x arch

* Remove helm & tiller from jailer for s390x

* Update shell image which supports s390x

* Add failure:ignore to drone s390x pipelines

This is to ensure that if any step of s390x pipeline fails, other steps or pipelines like manifest do not fail. So added ignore missing to manifest step pipeline

* Skip k3s

* Get k3s binaries from k3s-root for s390x as k3s is not available

* Remove duplicated code in Dockerfile

* Set K3S_BUILDER default value to rancher/k3s:v1.22.5-k3s1

+++ package/jailer.sh
@@ -58,13 +58,17 @@ cp -r -l /opt/drivers/management-state/bin /opt/jail/$NAME/var/lib/rancher/manag
 cp -l /usr/bin/rancher-machine /opt/jail/$NAME/usr/bin
 
 # Hard link helm_2 into the jail
-cp -l /usr/bin/rancher-helm /opt/jail/$NAME/usr/bin
+if [[ -f /usr/bin/rancher-helm ]]; then
+  cp -l /usr/bin/rancher-helm /opt/jail/$NAME/usr/bin
+fi
 
 # Hard link helm_3 into the jail
 cp -l /usr/bin/helm_v3 /opt/jail/$NAME/usr/bin
 
 # Hard link tiller into the jail
-cp -l /usr/bin/rancher-tiller /opt/jail/$NAME/usr/bin
+if [[ -f /usr/bin/rancher-tiler ]]; then
+  cp -l /usr/bin/rancher-tiller /opt/jail/$NAME/usr/bin
+fi
 
 # Hard link kustomize into the jail
 cp -l /usr/bin/kustomize /opt/jail/$NAME/usr/bin
